feat(tool): refactor aish-tool crate
- remove tool description from i18n crate

- add webfetch tool
