+++ crates/aish-i18n/locales/es-ES.yaml
@@ -353,25 +353,9 @@ security:
 
 tools:
   ask_user:
-    description: "Haz al usuario una pregunta de aclaración enfocada. Con solo prompt se usa texto libre; con options se ofrecen elecciones."
     unknown_kind: "Tipo desconocido: {kind}"
     runtime_not_configured: "el runtime de ask_user no está configurado"
     execute_failed: "ask_user falló: {error}"
-    param:
-      kind: "Tipo de interacción: text_input para texto libre, choice_or_text para opciones con entrada personalizada"
-      prompt: "La pregunta que se le hará al usuario."
-      options: "Opciones que se pueden ofrecer al usuario. Si faltan o están vacías, ask_user usa entrada de texto."
-      option_value: "Valor estable de la opción que se devuelve en los metadatos de la herramienta."
-      option_label: "Etiqueta visible para el usuario en la lista de opciones."
-      option_description: "Detalle opcional que se muestra debajo de la opción."
-      option_recommended: "Marca esta opción visiblemente como la recomendada."
-      title: "Título opcional para la pregunta"
-      default: "Valor predeterminado"
-      placeholder: "Texto de marcador"
-      allow_freeform_input: "Cuando hay opciones, permite elegir otra respuesta y escribirla. El valor predeterminado es true."
-      required: "Si el usuario debe proporcionar una respuesta (predeterminado: true)"
-      allow_cancel: "Si el usuario puede cancelar/omitir (predeterminado: true)"
-      min_length: "Longitud mínima para entrada de texto (predeterminado: 0)"
     validation:
       prompt_empty: "prompt no puede estar vacío"
       option_value_empty: "option value no puede estar vacío"
@@ -388,10 +372,6 @@ tools:
       cancelled: "El usuario canceló la pregunta."
 
   smart_log:
-    description: "Determina automáticamente las fuentes de registros, analiza los logs según la consulta y devuelve un resumen coincidente."
-    param:
-      query: "Tu consulta de diagnóstico, por ejemplo 'nginx está bien configurado' o 'hay inicios de sesión sospechosos'."
-      path: "Opcional. Ruta del log (archivo/directorio) o nombre de una unidad systemd. Si se omite, se inferirá a partir de la consulta."
     summary:
       file: "[ARCHIVO] {ident}: {matched}/{total} líneas coincidentes\nMuestra:\n{sample}"
       dir: "[DIR] {path}: {matched}/{total} líneas coincidentes\nMuestra:\n{sample}"
