+++ crates/aish-i18n/locales/ja-JP.yaml
@@ -353,25 +353,9 @@ security:
 
 tools:
   ask_user:
-    description: "ユーザーに焦点を絞った確認質問をします。prompt のみならテキスト入力、options があれば選択肢を提示します。"
     unknown_kind: "不明な種類です: {kind}"
     runtime_not_configured: "ask_user ランタイムが構成されていません"
     execute_failed: "ask_user に失敗しました: {error}"
-    param:
-      kind: "操作タイプ: text_input は自由入力、choice_or_text は選択肢とカスタム入力"
-      prompt: "ユーザーに尋ねる質問です。"
-      options: "ユーザーに提示する任意の選択肢です。省略または空の場合、ask_user はテキスト入力を使います。"
-      option_value: "ツールメタデータに返される安定した選択肢の値です。"
-      option_label: "選択リストに表示されるユーザー向けラベルです。"
-      option_description: "選択肢の下に表示される任意の補足説明です。"
-      option_recommended: "この選択肢を推奨として明示表示します。"
-      title: "質問の任意タイトル"
-      default: "既定値"
-      placeholder: "プレースホルダーテキスト"
-      allow_freeform_input: "選択肢がある場合にカスタム回答を許可します。既定値は true です。"
-      required: "ユーザーが回答を入力する必要があるかどうか（既定値: true）"
-      allow_cancel: "ユーザーがキャンセル/スキップできるかどうか（既定値: true）"
-      min_length: "テキスト入力の最小文字数（既定値: 0）"
     validation:
       prompt_empty: "prompt は空にできません"
       option_value_empty: "option value は空にできません"
@@ -388,10 +372,6 @@ tools:
       cancelled: "ユーザーが質問をキャンセルしました。"
 
   smart_log:
-    description: "ログソースを自動判定し、問い合わせに基づいてログを解析し、一致する要約を返します。"
-    param:
-      query: "診断したい内容。例: nginx は正しく設定されているか、怪しいログインはあるか。"
-      path: "任意。ログパス（ファイルまたはディレクトリ）または systemd ユニット名。省略時は問い合わせから推定します。"
     summary:
       file: "[FILE] {ident}: 一致 {matched}/{total} 行\nサンプル:\n{sample}"
       dir: "[DIR] {path}: 一致 {matched}/{total} 行\nサンプル:\n{sample}"
