+++ crates/aish-tools/src/read_file/read_file.rs
@@ -1,0 +1,252 @@
+use aish_i18n;
+use aish_llm::{Tool, ToolResult};
+
+use super::prompt;
+
+/// Read file content tool.
+pub struct ReadFileTool;
+
+impl Default for ReadFileTool {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ReadFileTool {
+    pub fn new() -> Self {
+        Self
+    }
+}
+
+impl Tool for ReadFileTool {
+    fn name(&self) -> &str {
+        "read_file"
+    }
+
+    fn description(&self) -> &str {
+        prompt::DESCRIPTION
+    }
+
+    fn parameters(&self) -> serde_json::Value {
+        prompt::parameters()
+    }
+
+    fn prompt(&self) -> &str {
+        prompt::PROMPT
+    }
+
+    fn execute(&self, args: serde_json::Value) -> ToolResult {
+        let path = match args.get("path").and_then(|p| p.as_str()) {
+            Some(p) => p,
+            None => return ToolResult::error(aish_i18n::t("tools.fs.read_file.missing_path")),
+        };
+
+        let raw_bytes = match std::fs::read(path) {
+            Ok(b) => b,
+            Err(e) => {
+                let mut args_map = std::collections::HashMap::new();
+                args_map.insert("path".to_string(), path.to_string());
+                args_map.insert("error".to_string(), e.to_string());
+                return ToolResult::error(aish_i18n::t_with_args(
+                    "tools.fs.read_file.read_failed",
+                    &args_map,
+                ));
+            }
+        };
+
+        const SIZE_LIMIT: usize = 32 * 1024;
+        if raw_bytes.len() > SIZE_LIMIT {
+            let mut args_map = std::collections::HashMap::new();
+            args_map.insert("path".to_string(), path.to_string());
+            args_map.insert("size".to_string(), raw_bytes.len().to_string());
+            args_map.insert("limit".to_string(), SIZE_LIMIT.to_string());
+            return ToolResult::error(aish_i18n::t_with_args(
+                "tools.fs.read_file.file_too_large",
+                &args_map,
+            ));
+        }
+
+        let content = match String::from_utf8(raw_bytes) {
+            Ok(s) => s,
+            Err(e) => {
+                let mut args_map = std::collections::HashMap::new();
+                args_map.insert("path".to_string(), path.to_string());
+                args_map.insert("error".to_string(), e.to_string());
+                return ToolResult::error(aish_i18n::t_with_args(
+                    "tools.fs.read_file.decode_failed",
+                    &args_map,
+                ));
+            }
+        };
+
+        let lines: Vec<&str> = content.lines().collect();
+        if lines.is_empty() {
+            return ToolResult::success(aish_i18n::t("tools.fs.read_file.empty_file"));
+        }
+
+        let offset = args.get("offset").and_then(|o| o.as_u64()).unwrap_or(0) as usize;
+        let limit = args
+            .get("limit")
+            .and_then(|l| l.as_u64())
+            .map(|l| l as usize);
+
+        if offset >= lines.len() {
+            let mut args_map = std::collections::HashMap::new();
+            args_map.insert("offset".to_string(), offset.to_string());
+            args_map.insert("length".to_string(), lines.len().to_string());
+            return ToolResult::error(aish_i18n::t_with_args(
+                "tools.fs.read_file.offset_exceeds_length",
+                &args_map,
+            ));
+        }
+
+        let selected: Vec<String> = if let Some(limit) = limit {
+            lines
+                .iter()
+                .skip(offset)
+                .take(limit)
+                .enumerate()
+                .map(|(i, line)| format!("{:>6}\t{}", offset + i + 1, line))
+                .collect()
+        } else {
+            lines
+                .iter()
+                .skip(offset)
+                .enumerate()
+                .map(|(i, line)| format!("{:>6}\t{}", offset + i + 1, line))
+                .collect()
+        };
+
+        ToolResult::success(selected.join("\n"))
+    }
+}
+
+/// Path-restricted wrapper around [`ReadFileTool`] for SSH sessions.
+pub struct SshReadFileTool {
+    inner: ReadFileTool,
+    offload_root: std::path::PathBuf,
+}
+
+impl SshReadFileTool {
+    pub fn new() -> Self {
+        let offload_root = std::env::temp_dir().join("aish-offload");
+        let canonical_root = std::fs::canonicalize(&offload_root).unwrap_or(offload_root);
+        Self {
+            inner: ReadFileTool::new(),
+            offload_root: canonical_root,
+        }
+    }
+}
+
+impl Tool for SshReadFileTool {
+    fn name(&self) -> &str {
+        self.inner.name()
+    }
+
+    fn description(&self) -> &str {
+        self.inner.description()
+    }
+
+    fn parameters(&self) -> serde_json::Value {
+        self.inner.parameters()
+    }
+
+    fn prompt(&self) -> &str {
+        self.inner.prompt()
+    }
+
+    fn execute(&self, args: serde_json::Value) -> ToolResult {
+        let path = match args.get("path").and_then(|p| p.as_str()) {
+            Some(p) => p,
+            None => return ToolResult::error(aish_i18n::t("tools.fs.read_file.missing_path")),
+        };
+        let canonical = match std::fs::canonicalize(path) {
+            Ok(c) => c,
+            Err(e) => {
+                let mut args_map = std::collections::HashMap::new();
+                args_map.insert("path".to_string(), path.to_string());
+                args_map.insert("error".to_string(), e.to_string());
+                return ToolResult::error(aish_i18n::t_with_args(
+                    "tools.fs.read_file.read_failed",
+                    &args_map,
+                ));
+            }
+        };
+        if !canonical.starts_with(&self.offload_root) {
+            return ToolResult::error("Access denied: path is not inside offload directory");
+        }
+        let mut safe_args = args;
+        if let Some(obj) = safe_args.as_object_mut() {
+            obj.insert(
+                "path".to_string(),
+                serde_json::Value::String(canonical.to_string_lossy().into_owned()),
+            );
+        }
+        self.inner.execute(safe_args)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use aish_llm::Tool;
+    use std::fs;
+
+    fn temp_dir() -> tempfile::TempDir {
+        tempfile::tempdir().expect("failed to create temp dir")
+    }
+
+    #[test]
+    fn test_read_file_with_line_numbers() {
+        let dir = temp_dir();
+        let file_path = dir.path().join("test.txt");
+        fs::write(&file_path, "hello\nworld\nfoo").unwrap();
+
+        let tool = ReadFileTool::new();
+        let result = tool.execute(serde_json::json!({
+            "path": file_path.to_str().unwrap()
+        }));
+
+        assert!(result.ok);
+        assert_eq!(result.output, "     1\thello\n     2\tworld\n     3\tfoo");
+    }
+
+    #[test]
+    fn test_read_file_with_offset() {
+        let dir = temp_dir();
+        let file_path = dir.path().join("test.txt");
+        fs::write(&file_path, "line1\nline2\nline3\nline4\nline5").unwrap();
+
+        let tool = ReadFileTool::new();
+        let result = tool.execute(serde_json::json!({
+            "path": file_path.to_str().unwrap(),
+            "offset": 2,
+            "limit": 2
+        }));
+
+        assert!(result.ok);
+        assert_eq!(result.output, "     3\tline3\n     4\tline4");
+    }
+
+    #[test]
+    fn test_read_file_size_limit() {
+        aish_i18n::set_locale("en-US");
+
+        let dir = temp_dir();
+        let file_path = dir.path().join("big.txt");
+        let big_content = "x".repeat(33 * 1024);
+        fs::write(&file_path, &big_content).unwrap();
+
+        let tool = ReadFileTool::new();
+        let result = tool.execute(serde_json::json!({
+            "path": file_path.to_str().unwrap()
+        }));
+
+        assert!(!result.ok);
+        assert!(
+            result.output.contains("limit") || result.output.contains("bytes"),
+            "Expected size limit error, got: {}",
+            result.output
+        );
+    }
+}
